extracting spy logic into its own file
